+++ .gitignore
@@ -23,6 +23,7 @@ bld/
 [Bb]in/
 [Oo]bj/
 [Ll]og/
+coverage*
 
 # Visual Studio 2015/2017 cache/options directory
 .vs/
