+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Change UseInvalidModelStateResponseFactory to SuppressInvalidModelStateResponseFactoryOverride
+Added support for .Net Core 3.0 (3.0.100-preview6-012264).
 
 ## Improvements/Features
-* [??](https://github.com/ofpinewood/http-exceptions/issues/??) ?? 
+* [#21](https://github.com/ofpinewood/http-exceptions/issues/21) .Net Core 3.0 support
