+++ test-netcoreapp3_0.ps1
@@ -1,0 +1,3 @@
+dotnet build --configuration Release
+dotnet test --configuration Release --framework netcoreapp3.0 --no-build /p:CollectCoverage=true /p:CoverletOutputFormat=cobertura /p:Include="[Opw.*]*" /p:Exclude="[*.Tests]*%2c[*.*Tests]*"
+PAUSE
