+++ tests/Opw.HttpExceptions.AspNetCore.IntegrationTests/NotUseInvalidModelStateResponseFactoryTests.cs
@@ -21,7 +21,7 @@ namespace Opw.HttpExceptions.AspNetCore
         public async Task PostProduct_Should_ReturnOk()
         {
             var product = new Product { Id = "1" };
-            var response = await _client.PostAsync("test/product", product.ToHttpContent());
+            var response = await _client.PostAsync("test/product", product.ToJsonContent());
 
             response.IsSuccessStatusCode.Should().BeTrue();
             response.StatusCode.Should().Be(HttpStatusCode.OK);
@@ -31,7 +31,7 @@ namespace Opw.HttpExceptions.AspNetCore
         public async Task PostProduct_Should_ReturnProblemDetails_UsingAspNetCoreDefaultImplementation()
         {
             var product = new Product();
-            var response = await _client.PostAsync("test/product", product.ToHttpContent());
+            var response = await _client.PostAsync("test/product", product.ToJsonContent());
 
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
             response.Content.Headers.ContentType.MediaType.Should().Be("application/problem+json");
@@ -44,7 +44,7 @@ namespace Opw.HttpExceptions.AspNetCore
             problemDetails.Type.Should().BeNull();
             problemDetails.Detail.Should().BeNull();
             problemDetails.Instance.Should().BeNull();
-            problemDetails.Extensions.Should().HaveCount(2);
+            problemDetails.Extensions.Should().HaveCountGreaterOrEqualTo(1);
             problemDetails.Extensions.ContainsKey("traceId").Should().BeTrue();
         }
     }
