+++ tests/Opw.HttpExceptions.AspNetCore.Tests/ProblemDetailsExtensions.cs
@@ -1,5 +1,6 @@
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json.Linq;
 using System.Net;
 
 namespace Opw.HttpExceptions.AspNetCore
@@ -27,6 +28,29 @@ namespace Opw.HttpExceptions.AspNetCore
             exceptionDetails.StackTrace.Should().NotBeNull();
 
             return exceptionDetails;
+        }
+
+        public static bool TryGetExceptionDetails(this ProblemDetails problemDetails, out ExceptionDetails exceptionDetails)
+        {
+            if (problemDetails.Extensions.TryGetValue(nameof(ExceptionDetails).ToCamelCase(), out var value))
+                return value.TryParseExceptionDetails(out exceptionDetails);
+
+            exceptionDetails = null;
+            return false;
+        }
+
+        internal static bool TryParseExceptionDetails(this object value, out ExceptionDetails exceptionDetails)
+        {
+            exceptionDetails = null;
+
+#pragma warning disable RCS1220 // Use pattern matching instead of combination of 'is' operator and cast operator.
+            if (value is ExceptionDetails)
+                exceptionDetails = (ExceptionDetails)value;
+            if (value is JToken)
+                exceptionDetails = ((JToken)value).ToObject<ExceptionDetails>();
+#pragma warning restore RCS1220 // Use pattern matching instead of combination of 'is' operator and cast operator.
+
+            return exceptionDetails != null;
         }
     }
 }
